transformer,generator,at_types.py: Expand eras['rules'] into policy_name and delta_seconds
There are 3 types of RULES: '-', 'hh:ss', and a string reference to a set of
RULE records (i.e. policy). Previously, this field was clobbered to be just
'-', ':', or a string reference. Now, this field is unchanged, which makes it easier
to generate notable comments referring to this field.

If the RULES is a policy name, then the string is extracted to a
era['policy_name'] field. This field is set to 'None' if RULES is either '-' or
'hh:mm', making it much easier to distinguish between the different types.